FasterPythonServices/requirements.txt
aiohttp==3.0.7
asyncpg==0.15.0
celery==4.1.0
uwsgi==2.0.15
flask==0.12.2
gunicorn==19.7.1
ipython==6.2.1
msgpack-python==0.4.8
pillow==4.3.0
psycopg2==2.7.3.2
python==3.6.4
pyzmq==16.0.3
redis==2.10.6
requests==2.18.4
pybloom_live==2.3.1
influxdb==5.0.0
